voice_quality_tool/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pyaudio>=0.2.11
librosa>=0.9.0
openai>=1.0.0
matplotlib>=3.3.0
dtw-python>=1.3.0  # 可选：用于DTW精对齐（或使用 fastdtw）
# fastdtw>=0.3.0  # 备选：如果dtw-python安装失败
